Do not always install udev file to the system
Currently the user is forced to have root permissions on the system when
using the `install` target. This can cause problems with tkDeps or Yocto
build or, more in general, with automatic build systems where usually
we don't have root permission and we would have at least the opportunity
to chose if we want to install something on the root system.
Make the udev install optional using a variable. Update the Readme with
the new option.

Signed-off-by: Luca Miccio <lucmiccio@gmail.com>